NodePool (untested) added
A node pool unifies multiple local nodes. The intention is providing
things like a bootstrap or drawing server an easy way to spawn its own
nodes for which they have special permission to send signals to. This
simplifies connecting special clients to the network greatly without
compromising the network principle of anonymity and
decentralization.strap or drawing server an easy way to spawn its own
nodes for which they have special permission to send signals to. This
simplifies connecting special clients to the network greatly without
compromising the network principle of anonymity and decentralization.