Rework housekeeping [Leak fixed! For real this time]
Combining the two downstream-removing functions to one seemed useful
because there was some work duplication going on.
The surprising consequence of this is change is the fix of the eternal
leak. Apparently the old functions did sometimes cancel a client but
keep the entry; it was periodically fed KeepAlive signals, occupied a
zombie spot in the DB, and did so until its private queue was full and
everything was doomed in general.