Fix 'too many USNs' problem
Having restarts during the initial launch of the bootstrap server's node
pool seemed to be pretty bad. I'm not sure what the exact reason for
this behaviour is, but I would assume that nodes don't time out fast
enough in this critical stage of the network creation, leaving orphan
connections open that block new connections from being formed. When the
timeouts finally kick in, everything dies.